// components/numberRange/numberRangeTargets.enum.ts
export enum NumberRangeTargets {
  root = 'root',
  label = 'label',
  input = 'input',
  helperText = 'helperText',
  errors = 'errors',
  inputPrefix = 'inputPrefix',
  inputSuffix = 'inputSuffix',
  readOnly = 'readOnly',
  tooltipButton = 'tooltipButton',
  tooltipPopover = 'tooltipPopover',
  innerWrapper = 'innerWrapper',
  innerWrapperChild = 'innerWrapperChild',
  minMaxRangeWrapper = 'minMaxRangeWrapper',
  minRange = 'minRange',
  maxRange = 'maxRange',
  displayValueWrapper = 'displayValueWrapper',
  displayValue = 'displayValue',
}